@@ -15,7 +15,7 @@
     "src"
   ],
   "engines": {
-    "node": ">=10"
+    "node": ">= 16.5.0"
   },
   "scripts": {
     "dev": "tsdx watch",
@@ -25,7 +25,8 @@
     "prepare": "tsdx build"
   },
   "dependencies": {
-    "@formatjs/cli": "5.0.2",
+    "@formatjs/cli": "^5.0.2",
+    "@formatjs/ts-transformer": "^3.9.9",
     "fast-glob": "^3.2.11",
     "minimist": "^1.2.6",
     "node-fetch": "2.6.7"
